Revert "修改数据类型double为float，K60只有单精度浮点运算单元，双精度是CPU模拟，速度慢"
This reverts commit 6c3a23e9ecedb58511390349a260dec31c1d7545.

+++ App/Inc/motor.h
@@ -8,8 +8,8 @@ extern void Motor_PID();
 extern void Get_Motor_Speed_Init();
 extern void Get_Motor_Speed();
 extern void FuzzyPID();
-extern float FuzzyKp(int16 e, float ec);
-extern float FuzzyKi(int e, float ec);
-extern float FuzzyKd(int e, float ec);
+extern double FuzzyKp(int16 e, double ec);
+extern double FuzzyKi(int e, double ec);
+extern double FuzzyKd(int e, double ec);
 
 #endif //__MOTOR_H__
